Moved drag functionality to useDragPlacement hook
